Move registry into a dedicated package
