+++ pkg/validation/linters.go
@@ -19,7 +19,7 @@ import (
 	"fmt"
 	"strings"
 
-	"sigs.k8s.io/kube-api-linter/pkg/analysis"
+	"sigs.k8s.io/kube-api-linter/pkg/analysis/registry"
 	"sigs.k8s.io/kube-api-linter/pkg/config"
 
 	"k8s.io/apimachinery/pkg/util/sets"
@@ -40,8 +40,8 @@ func ValidateLinters(l config.Linters, fldPath *field.Path) field.ErrorList {
 		fieldErrors = append(fieldErrors, field.Invalid(enablePath, l.Enable, "values in 'enable' must be unique"))
 	case enable.Has(config.Wildcard) && enable.Len() != 1:
 		fieldErrors = append(fieldErrors, field.Invalid(enablePath, l.Enable, "wildcard ('*') must not be specified with other values"))
-	case !enable.Has(config.Wildcard) && enable.Difference(analysis.DefaultRegistry().AllLinters()).Len() > 0:
-		fieldErrors = append(fieldErrors, field.Invalid(enablePath, l.Enable, fmt.Sprintf("unknown linters: %s", strings.Join(enable.Difference(analysis.DefaultRegistry().AllLinters()).UnsortedList(), ","))))
+	case !enable.Has(config.Wildcard) && enable.Difference(registry.DefaultRegistry().AllLinters()).Len() > 0:
+		fieldErrors = append(fieldErrors, field.Invalid(enablePath, l.Enable, fmt.Sprintf("unknown linters: %s", strings.Join(enable.Difference(registry.DefaultRegistry().AllLinters()).UnsortedList(), ","))))
 	}
 
 	disable := sets.New(l.Disable...)
@@ -52,8 +52,8 @@ func ValidateLinters(l config.Linters, fldPath *field.Path) field.ErrorList {
 		fieldErrors = append(fieldErrors, field.Invalid(disablePath, l.Disable, "values in 'disable' must be unique"))
 	case disable.Has(config.Wildcard) && disable.Len() != 1:
 		fieldErrors = append(fieldErrors, field.Invalid(disablePath, l.Disable, "wildcard ('*') must not be specified with other values"))
-	case !disable.Has(config.Wildcard) && disable.Difference(analysis.DefaultRegistry().AllLinters()).Len() > 0:
-		fieldErrors = append(fieldErrors, field.Invalid(disablePath, l.Disable, fmt.Sprintf("unknown linters: %s", strings.Join(disable.Difference(analysis.DefaultRegistry().AllLinters()).UnsortedList(), ","))))
+	case !disable.Has(config.Wildcard) && disable.Difference(registry.DefaultRegistry().AllLinters()).Len() > 0:
+		fieldErrors = append(fieldErrors, field.Invalid(disablePath, l.Disable, fmt.Sprintf("unknown linters: %s", strings.Join(disable.Difference(registry.DefaultRegistry().AllLinters()).UnsortedList(), ","))))
 	}
 
 	if enable.Intersection(disable).Len() > 0 {
